fix(iOS): download beauty resources from Artifactory
- download FaceUnity and SenseTime resources from Artifactory
- remove Jenkins beauty resource URL dependencies
- validate required beauty resource files after extraction

+++ iOS/APIExample/cloud_build.sh
@@ -102,15 +102,19 @@ echo "-----------------------------------"
 cd ${PROJECT_PATH}
 
 #下载美颜资源
-echo "start download sense resource : $sense_lib"
-curl -L -O "$sense_lib"
-unzip -o vender_sense_iOS.zip
+echo "start download sense resource from Artifactory"
+curl --fail --location -H "X-JFrog-Art-Api:${JFROG_API_KEY}" -o vender_sense_iOS.zip "https://artifactory-api.bj2.agoralab.co/artifactory/qa_test_data/beauty/vender_sense_iOS.zip" || exit 1
+rm -rf SenseLib
+unzip -q vender_sense_iOS.zip || exit 1
 rm -f vender_sense_iOS.zip
+test -f SenseLib/SENSEME.lic || exit 1
 
-echo "start download fu resource : $fu_lib"
-curl -L -O "$fu_lib"
-unzip -o vender_fu_iOS.zip
+echo "start download fu resource from Artifactory"
+curl --fail --location -H "X-JFrog-Art-Api:${JFROG_API_KEY}" -o vender_fu_iOS.zip "https://artifactory-api.bj2.agoralab.co/artifactory/qa_test_data/beauty/vender_fu_iOS.zip" || exit 1
+rm -rf FULib
+unzip -q vender_fu_iOS.zip || exit 1
 rm -f vender_fu_iOS.zip
+test -f FULib/authpack.h || exit 1
 
 #打开第三方播放器配置
 sed -i -e "s#\#  pod 'ijkplayer'#  pod 'ijkplayer'#g" Podfile
